Move MD5 code into a separate 'md5' component
