+++ esphome/components/md5/__init__.py
@@ -1,0 +1,5 @@
+import esphome.config_validation as cv
+
+CODEOWNERS = ["@esphome/core"]
+
+CONFIG_SCHEMA = cv.Schema({})
